requirements.txt
python-dotenv
pandas 
numpy
scipy
scikit-learn
pymongo
certifi
pymongo[srv]==3.6
pyaml

##-e .
